Reverted breaking things and split up huge cs file into classes
